Implemented lexer

+++ jomad/src/main/kotlin/lexer/Token.kt
@@ -1,7 +1,7 @@
 package lexer
 
 sealed interface Token {
-    data class NumberLiteral(val value: Int) : Token
+    data class NumberLiteral(val value: Double) : Token
     data class StringLiteral(val value: String) : Token
     data class BooleanLiteral(val value: Boolean) : Token
     data class SymbolLiteral(val value: String) : Token
